PycharmProjects/bitr/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
yookassa==3.0.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6

